+++ algorithms/selectionSort.cpp
@@ -1,6 +1,6 @@
 
 // Comparation
-void selectionSort_compare(int arr[], int n, int& count_compare) {
+void selectionSort_compare(int arr[], int n, long long int &count_compare) {
     count_compare = 0;
 
     for (int i = 0; ++count_compare && i < n; i++) {
